requirements.txt
pandas
numpy
matplotlib
seaborn
scipy==1.7.3
pymc3==3.11.5
arviz
jupyter
notebook
scikit-learn
Flask==2.3.3

